refactor(server): extract analytics types in its own file as per convention
