+++ packages/amplication-server/src/core/account/account.service.ts
@@ -2,11 +2,9 @@ import { Injectable } from "@nestjs/common";
 import { Account, Prisma } from "../../prisma";
 import { PrismaService } from "../../prisma/prisma.service";
 import { Workspace } from "../../models";
-import {
-  SegmentAnalyticsService,
-  EnumEventType,
-  IdentifyData,
-} from "../../services/segmentAnalytics/segmentAnalytics.service";
+import { SegmentAnalyticsService } from "../../services/segmentAnalytics/segmentAnalytics.service";
+import { EnumEventType } from "../../services/segmentAnalytics/segmentAnalytics.types";
+import { IdentifyData } from "../../services/segmentAnalytics/segmentAnalytics.types";
 
 @Injectable()
 export class AccountService {
