+++ packages/amplication-server/src/services/segmentAnalytics/segmentAnalytics.service.ts
@@ -2,69 +2,7 @@ import { Injectable, Inject } from "@nestjs/common";
 import Analytics from "analytics-node";
 import { SegmentAnalyticsOptions } from "./segmentAnalytics.interfaces";
 import { RequestContext } from "nestjs-request-context";
-
-export enum EnumEventType {
-  Signup = "Signup",
-  WorkspacePlanUpgradeRequest = "WorkspacePlanUpgradeRequest",
-  WorkspacePlanUpgradeCompleted = "WorkspacePlanUpgradeCompleted",
-  WorkspacePlanDowngradeRequest = "WorkspacePlanDowngradeRequest",
-  CommitCreate = "commit",
-  WorkspaceSelected = "selectWorkspace",
-  GitHubAuthResourceComplete = "completeAuthResourceWithGitHub",
-  ServiceWizardServiceGenerated = "ServiceWizard_ServiceGenerated",
-  SubscriptionLimitPassed = "SubscriptionLimitPassed",
-  EntityCreate = "createEntity",
-  EntityUpdate = "updateEntity",
-  EntityFieldCreate = "createEntityField",
-  EntityFieldUpdate = "updateEntityField",
-  EntityFieldFromImportPrismaSchemaCreate = "EntityFieldFromImportPrismaSchemaCreate",
-  PluginInstall = "installPlugin",
-  PluginUpdate = "updatePlugin",
-  DemoRepoCreate = "CreateDemoRepo",
-  InvitationAcceptance = "invitationAcceptance",
-
-  //Import Prisma Schema
-  ImportPrismaSchemaStart = "importPrismaSchemaStart",
-  ImportPrismaSchemaError = "importPrismaSchemaError",
-  ImportPrismaSchemaCompleted = "importPrismaSchemaCompleted",
-
-  GitSyncError = "gitSyncError",
-  CodeGenerationError = "codeGenerationError",
-
-  CodeGeneratorVersionUpdate = "codeGeneratorVersionUpdate",
-
-  RedeemCoupon = "RedeemCoupon",
-
-  // break the monolith
-  ArchitectureRedesignStartRedesign = " architectureRedesign_StartRedesign",
-  ArchitectureRedesignApply = "architectureRedesign__Apply",
-  ArchitectureRedesignStartBreakTheMonolith = "architectureRedesign__StartBreakTheMonolith",
-}
-
-export type IdentifyData = {
-  userId: string;
-  firstName: string;
-  lastName: string;
-  email: string;
-  createdAt: Date;
-};
-
-export type TrackData = {
-  userId: string;
-  event: EnumEventType;
-  properties?:
-    | {
-        [key: string]: unknown;
-      }
-    | undefined;
-  context?: {
-    traits?: IdentifyData;
-    amplication?: {
-      analyticsSessionId?: string;
-    };
-  };
-};
-
+import { IdentifyData, TrackData } from "./segmentAnalytics.types";
 @Injectable()
 export class SegmentAnalyticsService {
   private analytics: Analytics;
